Fix bytewise parameter write acknowledgements
When QGC writes a parameter, it waits for the vehicle to echo the new
value in PARAM_VALUE. The acknowledgement is accepted only when both the
value and the QVariant type match QGC's stored Fact value.

For 8- & 16-bit parameters, FactMetaData stores the value in a QVariant:
Qt Int (32-bit) for signed values and 32-bit Qt UInt (32-bit) for
unsigned values. This preserves the parameter's signedness in QGC's
internal representation.

The `_mavlinkParamUnionToVariant` method passed `uint8_t` and `uint16_t`
directly to QVariant. C++ selected QVariant's `int` constructor, so an
unsigned PARAM_VALUE was decoded as signed Qt Int. The type check then
rejected the otherwise valid acknowledgement and the write timed out.
Decode narrow values into QGC's canonical Int or UInt representations.

The bug affects only UINT8 and UINT16 parameters; the other supported
types coincidentally work. It flew below the radar because ArduPilot
uses float-based parameter encoding, while PX4 does not expose UINT8 or
UINT16 parameters for which the problem exhibits itself.

+++ src/FactSystem/ParameterManager.cc
@@ -489,16 +489,16 @@ bool ParameterManager::_mavlinkParamUnionToVariant(const mavlink_param_union_t &
         outValue = QVariant(paramUnion.param_float);
         return true;
     case MAV_PARAM_TYPE_UINT8:
-        outValue = QVariant(paramUnion.param_uint8);
+        outValue = QVariant(static_cast<quint32>(paramUnion.param_uint8));
         return true;
     case MAV_PARAM_TYPE_INT8:
-        outValue = QVariant(paramUnion.param_int8);
+        outValue = QVariant(static_cast<qint32>(paramUnion.param_int8));
         return true;
     case MAV_PARAM_TYPE_UINT16:
-        outValue = QVariant(paramUnion.param_uint16);
+        outValue = QVariant(static_cast<quint32>(paramUnion.param_uint16));
         return true;
     case MAV_PARAM_TYPE_INT16:
-        outValue = QVariant(paramUnion.param_int16);
+        outValue = QVariant(static_cast<qint32>(paramUnion.param_int16));
         return true;
     case MAV_PARAM_TYPE_UINT32:
         outValue = QVariant(paramUnion.param_uint32);
